move export message tree models to oasst-shared
